// types.ts
export type AirportCode = 'ALL' | 'CGK' | 'DPS' | 'KNO' | 'SUB' | 'UPG' | 'BPN' | 'YIA' | 'DJJ' | string;

export type Page = 'PREDICTOR' | 'EXECUTIVE' | 'EVENT';

export interface DailyData {
  date: string; // YYYY-MM-DD
  dayName: string; // Senin, Selasa, etc.
  passengers: number;
  flights: number;
  isPeakDeparture?: boolean;
  isPeakReturn?: boolean;
  isHoliday?: boolean; // Natal or Tahun Baru
  description?: string;
}